+++ packages/web-sdk/src/transports/otlp/otel-payload/OtelPayload.ts
@@ -1,5 +1,5 @@
 import { Meta, TransportItem, TransportItemType } from '@grafana/faro-core';
-import { getResourceLogPayload, getScopeLog } from './transform/transforms';
+import { getResourceLogPayload, getScopeLog } from './transform/transform';
 import { internalLogger } from '../otlpPayloadLogger';
 
 import type { ResourceLogPayload } from './transform';
