+++ packages/web-sdk/src/transports/otlp/otel-payload/transform/getEventLogRecord.test.ts
@@ -1,5 +1,5 @@
 import { EventEvent, TransportItem, TransportItemType } from '@grafana/faro-core';
-import { getScopeLog } from './transforms';
+import { getScopeLog } from './transform';
 
 const item: TransportItem<EventEvent> = {
   type: TransportItemType.EVENT,
