+++ packages/web-sdk/src/transports/otlp/otel-payload/transform/transform.ts
@@ -14,7 +14,7 @@ import {
   TelemetrySdkLanguageValues,
 } from '@opentelemetry/semantic-conventions';
 import { internalLogger } from '../../otlpPayloadLogger';
-import { Attribute, isAttribute, toAttribute, toAttributeValue } from '../attributes';
+import { Attribute, isAttribute, toAttribute, toAttributeValue } from '../attribute';
 import type {
   LogTransportItem,
   ResourcePayload,
@@ -53,8 +53,9 @@ function getResource(transportItem: LogTransportItem): Readonly<ResourcePayload>
   return {
     attributes: [
       toAttribute(SemanticBrowserAttributes.BROWSER_MOBILE, browser?.mobile),
-      toAttribute(SemanticBrowserAttributes.BROWSER_USER_AGENT, undefined),
-      // toAttribute(SemanticBrowserAttributes.BROWSER_PLATFORM, browser?.os),
+      toAttribute(SemanticBrowserAttributes.BROWSER_USER_AGENT, browser?.userAgent),
+      toAttribute(SemanticBrowserAttributes.BROWSER_LANGUAGE, browser?.language),
+      toAttribute('browser.os', browser?.os),
       // toAttribute(SemanticBrowserAttributes.BROWSER_BRANDS, browser?.brands),
       toAttribute('browser.name', browser?.name),
       toAttribute('browser.version', browser?.version),
